+++ plugins/org.ietr.preesm.pimm.algorithm/src/org/ietr/preesm/pimm/algorithm/pimmoptims/PiMMOptimization.java
@@ -1,0 +1,22 @@
+/**
+ * 
+ */
+package org.ietr.preesm.pimm.algorithm.pimmoptims;
+
+import org.ietr.preesm.experiment.model.pimm.PiGraph;
+
+/**
+ * @author farresti
+ *
+ */
+public interface PiMMOptimization {
+
+  /**
+   * Perform graph optimizations on a given PiGraph
+   * 
+   * @param graph
+   *          the graph to optimize
+   */
+  boolean optimize(final PiGraph graph);
+
+}